Add keybinding for level 1 gnus start or switch
Create a function that always fetches a specified level even if gnus is
already started. This is different from the behavior of both `gnus' and
`gnus-other-frame' (although I might be missing a gnus function that
already does this). I'm fine frequently fetching level 1 because I
reserve this for mail groups (which are local).