Rework org agenda fullscreen advice
With previous advice, window configuration would not be restored
correctly if the agenda buffer was refreshed. This can be avoided if
restoring the window configuration is left to org mode (with
`org-agenda-restore-windows-after-quit'). There isn't a fullscreen
option in `org-agenda-window-setup', so `org-agenda-list' still needs to
be advised to delete the other windows.