Fix projectile switch and find file keybinding
"j" was previously unbound in the projectile keymap, but commit bebbab
added assigned it to a function to find project tags. I don't currently
use tags, so I want to keep my switch and find file keybinding. Now I
need to override the binding in the projectile keymap instead of using a
global keybinding.