Turn ido-vertical-mode back on
When I set up ace-link (809107e), I disabled ido-vertical-mode because I
was noticing that if it was activated, some of the links at the bottom
were being cut off. I can't reproduce this now (despite there not being
any changes in ido-vertical-mode, ace-jump, or ace-link to fix it). I'm
turning it back on and will see if the issue pops up again.