Add file-map prefix command
This prefix will be used for find-file functions that don't already have
a shorter keybinding.

This commit adds two functions to the map, `view-file' and
`magit-find-file', which both have an other-window analog in
`ctl-x-4-map'.

View mode can be accessed using ',v' right after finding a file, so the
`view-file' binding might not be too useful.

`magit-find-file' was in my magit-specific map, but it makes sense to
have it here because it's useful in buffers outside the magit status
buffer.