Add display-only version of git-rebase-show-commit
Bind this version to SPC.  This is more consistent with behavior in the
log buffer, where RET shows the commit and switches to that buffer and
SPC just displays the commit.

However, the SPC variants still differ in their scrolling behavior.  In
the log buffer, SPC (magit-diff-show-or-scroll-up) will scroll if the
buffer is already displayed, but km/git-rebase-show-commit doesn't
support this.  While it'd be nice to use magit-diff-show-or-scroll-up
directly, this isn't easy to do because magit-diff-show-or-scroll-up
doesn't take a revision argument (it pulls it directly from the magit
section).