Rework key chords
- Use letter pairs that are rare in both directions.  For really
  common commands, try to keep my fingers on home row.

- Remove some key chords for less frequent commands.  For Projectile,
  just bind switching projects and the commander because most other
  things are available with one more key from the commander.

My strategy for key chord bindings was to lead with ',' or ';' and
then follow with any letter. The idea was that, with the languages and
coding conventions I was using at the time, when I inserted a comma or
semicolon, they were usually followed by a non-letter (typically a
space or a new line).  But, this is actually not a good way to go
about this because Key chord mode accepts the two-letter combination
in either order.  In practice, I didn't end up unintentionally calling
a command too often, but it did happen enough to be slightly annoying.
John Cook recently posted a table [1] of bigram frequencies, which is
a good starting point for updating my key chords.

[1]: http://www.johndcook.com/blog/2015/02/01/rare-bigrams/