Add projectile-project-save-thing
Add commands to save and restore a buffer, file, or window configuration
for a project.

Projectile supports restoring window configuration with
persp-projectile.el.  (A recent change to Projectile [1] removed
projectile-remember-window-configs in favor of using perspective.el.)
However, I want to explicitly choose which (if any) window configuration
is saved for a project and also want the ability to save buffers or
files.

[1] c94c5f5ad385f58682c2ac3414066ec24dd75abe