Add delete-trailing-whitespace to cleanup function
With 5f4f111cd, I replaced by old buffer cleanup functions with
whitespace-mode. This resulted in similar behavior, but empty lines at
the end of the buffer were no longer deleted.  The 'empty' option for
whitespace-style will do this, but it also deletes empty lines at the
beginning of the buffer, which I don't want.

Add delete-trailing-whitespace to km/cleanup-buffer so that trailing
whitespace is removed.