Give in to Transient Mark mode
I tend to use C-SPC for pushing a position to jump back to later much
more than I use it to highlight a region, so I prefer to disable
Transient Mark mode and use C-SPC C-SPC to temporarily activate it.

However, it seems that lispy depends on this being on for its mark
commands to be useful.  So try to get used to C-SPC C-SPC to set a
mark without activating.

If I can't get used to that, I can try something similar to this:
http://www.masteringemacs.org/article/fixing-mark-commands-transient-mark-mode