Rewrite org-clone-and-shift-by-repeater
As of Org mode commit 329683861cc85aa,
org-clone-subtree-with-time-shift accepts a value of 0, which allows a
repeating subtree to be replaced with a non-repeating subtree and a
shifted, repeating subtree.

Rewrite org-clone-and-shift-by-repeater to use the new
org-clone-subtree-with-time-shift.  As an added bonus, this means that
org-clone-and-shift-by-repeater now supports scheduled items and time
stamps, not just deadlines