Remove -i flag from shell-command-switch
I was using this to get aliases, but this was a bad idea because it can
make internal commands that rely on this slow, especially if the
repeatedly spawn processes (I noticed this with magit-patch-id).

I will probably get annoyed soon by the lack of aliases with some
commands (e.g., compile) and will let-bind these variables around those
specific commands.