git-rebase-fixup-duplicates: Use marker
If only an integer is used, the end value used for the search is no
longer accurate because more characters are inserted for each line
where "fixup" or "squash" is substituted for "pick".  This discrepancy
is unlikely to have an observable effect when git-rebase-show-instructions
is non-nil because the end value includes many characters beyond the
last commit line.  But when git-rebase-show-instructions is nil, this
can easily result in the last commit line not matching the regexp
because the end value has shifted to the middle of the line.