diff-review: Fix function for comment movement
Rework km/diff-review-next-comment's move procedure to avoid an error
when the buffer doesn't contain any comments and to short-circuit the
recursion when the search function fails.  This also introduces a
change in behavior when N exceeds the number of comment blocks
before/after point: instead of always positioning point at the start
of the block, point is positioned at the end of the block for forward
movement and the start of the block for backward movement.